+++ src/main/java/com/jcloisterzone/Expansion.java
@@ -80,7 +80,7 @@ public class Expansion {
     @NotImplemented public static Expansion SCHOOL = new Expansion("SCHOOL", "SC", _tr("The School"), ExpansionType.PROMO);
     @NotImplemented public static Expansion LA_PORXADA = new Expansion("LA_PORXADA", "PX", _tr("La Porxada"), ExpansionType.PROMO);
     public static Expansion RUSSIAN_PROMOS = new Expansion("RUSSIAN_PROMOS", "RP", _tr("Russian Promos"), new Class[] { YagaCapability.class }, ExpansionType.PROMO);
-    @NotImplemented public static Expansion DARMSTADT_PROMO = new Expansion("DARMSTADT_PROMO", "DP", _tr("Darmstadt Promo"), ExpansionType.PROMO);
+    public static Expansion DARMSTADT_PROMO = new Expansion("DARMSTADT_PROMO", "DA", _tr("Darmstadt Promo"), new Class[] { ChurchCapability.class }, ExpansionType.PROMO);
     public static Expansion LABYRINTH = new Expansion("LABYRINTH", "LA", _tr("Labyrinth"), new Class[] { LabyrinthCapability.class }, ExpansionType.PROMO);
 
     @NotImplemented public static Expansion _MISSING_PLACEHOLDER = new Expansion("(Unknown expansion)", "??", _tr("(Unknown expansion)"), ExpansionType.UNKNOWN);
@@ -92,7 +92,7 @@ public class Expansion {
         KING_AND_ROBBER_BARON, RIVER, RIVER_II, CATHARS, BESIEGERS, COUNT, GQ11, CULT, TUNNEL,
         CORN_CIRCLES, PHANTOM, FESTIVAL, LITTLE_BUILDINGS, WIND_ROSE, GERMAN_MONASTERIES,
         FLIER, FERRIES, GOLDMINES, MAGE_AND_WITCH, CORN_CIRCLES_II,
-        RUSSIAN_PROMOS, LABYRINTH
+        RUSSIAN_PROMOS, DARMSTADT_PROMO, LABYRINTH
     );
 
     private final String name;