+++ src/main/java/com/jcloisterzone/feature/City.java
@@ -25,23 +25,26 @@ public class City extends CompletableFeature<City> {
 
     private final Set<Tuple2<ShortEdge, FeaturePointer>> multiEdges; // HS.CC!.v abstraction, multiple cities can connect to same edge
     private final int pennants;
+    private final int extraPoints;
     private final Map<TradeGoods, Integer> tradeGoods;
     private final boolean besieged, cathedral, princess, castleBase;
 
 
-    public City(List<FeaturePointer> places, Set<Edge> openEdges, int pennants) {
-        this(places, openEdges, HashSet.empty(), HashSet.empty(), pennants, HashMap.empty(), false, false, false, false);
+    public City(List<FeaturePointer> places, Set<Edge> openEdges, int pennants, int extraPoints) {
+        this(places, openEdges, HashSet.empty(), HashSet.empty(), pennants, extraPoints, HashMap.empty(), false, false, false, false);
     }
 
     public City(List<FeaturePointer> places,
             Set<Edge> openEdges, Set<FeaturePointer> neighboring,
             Set<Tuple2<ShortEdge, FeaturePointer>> multiEdges,
             int pennants,
+            int extraPoints,
             Map<TradeGoods, Integer> tradeGoods, boolean besieged, boolean cathedral, boolean princess,
             boolean castleBase) {
         super(places, openEdges, neighboring);
         this.multiEdges = multiEdges;
         this.pennants = pennants;
+        this.extraPoints = extraPoints;
         this.tradeGoods = tradeGoods;
         this.besieged = besieged;
         this.cathedral = cathedral;
@@ -58,6 +61,7 @@ public class City extends CompletableFeature<City> {
             mergeNeighboring(city),
             mergeMultiEdges(city),
             pennants + city.pennants,
+            extraPoints + city.extraPoints,
             mergeTradeGoods(city),
             besieged || city.besieged,
             cathedral || city.cathedral,
@@ -74,6 +78,7 @@ public class City extends CompletableFeature<City> {
             neighboring,
             multiEdges.filter(me -> !me._1.equals(edge)),
             pennants,
+            extraPoints,
             tradeGoods,
             besieged ,
             cathedral,
@@ -90,6 +95,7 @@ public class City extends CompletableFeature<City> {
             neighboring,
             multiEdges,
             pennants,
+            extraPoints,
             tradeGoods,
             besieged ,
             cathedral,
@@ -105,7 +111,7 @@ public class City extends CompletableFeature<City> {
             placeOnBoardEdges(pos, rot),
             placeOnBoardNeighboring(pos, rot),
             placeOnBoardMultiEdges(pos, rot),
-            pennants, tradeGoods, besieged, cathedral, princess, castleBase
+            pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase
         );
     }
 
@@ -115,7 +121,7 @@ public class City extends CompletableFeature<City> {
 
     public City setMultiEdges(Set<Tuple2<ShortEdge, FeaturePointer>> multiEdges) {
         if (this.multiEdges == multiEdges) return this;
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     public Set<Tuple2<ShortEdge, FeaturePointer>> getMultiEdges() {
@@ -125,7 +131,7 @@ public class City extends CompletableFeature<City> {
     @Override
     public City setNeighboring(Set<FeaturePointer> neighboring) {
         if (this.neighboring == neighboring) return this;
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     public boolean isBesieged() {
@@ -134,7 +140,7 @@ public class City extends CompletableFeature<City> {
 
     public City setBesieged(boolean besieged) {
         if (this.besieged == besieged) return this;
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     public boolean isCathedral() {
@@ -143,7 +149,7 @@ public class City extends CompletableFeature<City> {
 
     public City setCathedral(boolean cathedral) {
         if (this.cathedral == cathedral) return this;
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     public boolean isPrincess() {
@@ -152,7 +158,7 @@ public class City extends CompletableFeature<City> {
 
     public City setPrincess(boolean princess) {
         if (this.princess == princess) return this;
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     public boolean isCastleBase() {
@@ -161,19 +167,21 @@ public class City extends CompletableFeature<City> {
 
     public City setCastleBase(boolean castleBase) {
         if (this.castleBase == castleBase) return this;
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     public int getPennants() {
         return pennants;
     }
 
+    public int getExtraPoints() { return extraPoints; }
+
     public Map<TradeGoods, Integer> getTradeGoods() {
         return tradeGoods;
     }
 
     public City setTradeGoods(Map<TradeGoods, Integer> tradeGoods) {
-        return new City(places, openEdges, neighboring, multiEdges, pennants, tradeGoods, besieged, cathedral, princess, castleBase);
+        return new City(places, openEdges, neighboring, multiEdges, pennants, extraPoints, tradeGoods, besieged, cathedral, princess, castleBase);
     }
 
     private int getBasePoints(GameState state, boolean completed) {
@@ -197,7 +205,7 @@ public class City extends CompletableFeature<City> {
                 pointsPerUnit--;
             }
         }
-        return pointsPerUnit * (tileCount + pennants);
+        return pointsPerUnit * (tileCount + pennants) + extraPoints;
     }
 
     @Override