+++ src/main/java/com/jcloisterzone/feature/Cloister.java
@@ -26,24 +26,26 @@ public class Cloister extends TileFeature implements Scoreable, CloisterLike {
 
     protected final Set<FeaturePointer> neighboring; //for wagon move
 
-    protected final boolean shrine;
-    protected final boolean monastery;
+    protected final boolean shrine; // Cult expansion
+    protected final boolean monastery; // Monasteries expansion
+    protected final boolean church; // Darmstadt promo expansion
 
     public Cloister() {
-        this(INITIAL_PLACE, HashSet.empty(), false, false);
+        this(INITIAL_PLACE, HashSet.empty(), false, false, false);
     }
 
-    public Cloister(List<FeaturePointer> places, Set<FeaturePointer> neighboring, boolean shrine, boolean monastery) {
+    public Cloister(List<FeaturePointer> places, Set<FeaturePointer> neighboring, boolean shrine, boolean monastery, boolean church) {
         super(places);
         this.neighboring = neighboring;
         this.shrine = shrine;
         this.monastery = monastery;
+        this.church = church;
     }
 
     @Override
     public Cloister setNeighboring(Set<FeaturePointer> neighboring) {
         if (this.neighboring == neighboring) return this;
-        return new Cloister(places, neighboring, shrine, monastery);
+        return new Cloister(places, neighboring, shrine, monastery, church);
     }
 
     @Override
@@ -53,7 +55,7 @@ public class Cloister extends TileFeature implements Scoreable, CloisterLike {
 
     @Override
     public Feature placeOnBoard(Position pos, Rotation rot) {
-        return new Cloister(placeOnBoardPlaces(pos, rot), placeOnBoardNeighboring(pos, rot), shrine, monastery);
+        return new Cloister(placeOnBoardPlaces(pos, rot), placeOnBoardNeighboring(pos, rot), shrine, monastery, church);
     }
 
     public boolean isShrine() {
@@ -62,7 +64,7 @@ public class Cloister extends TileFeature implements Scoreable, CloisterLike {
 
     public Cloister setShrine(boolean shrine) {
         if (this.shrine == shrine) return this;
-        return new Cloister(places, neighboring, shrine, monastery);
+        return new Cloister(places, neighboring, shrine, monastery, church);
     }
 
     public boolean isMonastery() {
@@ -71,7 +73,16 @@ public class Cloister extends TileFeature implements Scoreable, CloisterLike {
 
     public Cloister setMonastery(boolean monastery) {
         if (this.monastery == monastery) return this;
-        return new Cloister(places, neighboring, shrine, monastery);
+        return new Cloister(places, neighboring, shrine, monastery, church);
+    }
+
+    public boolean isChurch() {
+        return church;
+    }
+
+    public Cloister setChurch(boolean church) {
+        if (this.church == church) return this;
+        return new Cloister(places, neighboring, shrine, monastery, church);
     }
 
     @Override